security-agent/requirements.txt
# Bridge Security Agent - Python Dependencies
# AI-powered code security scanner and fixer

# Google Gemini AI SDK
google-generativeai>=0.5.0

# Git operations
GitPython>=3.1.0

# HTTP requests for API calls
requests>=2.31.0

# JSON Schema validation
jsonschema>=4.20.0

# Type hints support
typing-extensions>=4.9.0
